requirements.txt
absl-py==0.3.0
asn1crypto==0.24.0
astetik==1.9.8
astor==0.7.1
astroid==2.1.0
attrs==18.2.0
audioread==2.1.6
autopep8==1.4.4
backcall==0.1.0
bandit==1.6.0
bleach==1.5.0
certifi==2019.6.16
cffi==1.11.5
chances==0.1.6
chardet==3.0.4
codacy-coverage==1.3.11
colorama==0.4.1
coverage==4.5.3
cryptography==2.3.1
cycler==0.10.0
decorator==4.3.0
defusedxml==0.6.0
entrypoints==0.3
eradicate==1.0
flake8==3.7.7
flake8-annotations-complexity==0.0.2
flake8-bandit==2.1.1
flake8-broken-line==0.1.0
flake8-bugbear==18.8.0
flake8-builtins==1.4.1
flake8-coding==1.3.1
flake8-commas==2.0.0
flake8-comprehensions==2.1.0
flake8-debugger==3.1.0
flake8-docstrings==1.3.0
flake8-eradicate==0.2.0
flake8-isort==2.7.0
flake8-logging-format==0.6.0
flake8-pep3101==1.2.1
flake8-polyfill==1.0.2
flake8-print==3.1.0
flake8-quotes==1.0.0
flake8-string-format==0.2.3
future==0.17.0
gast==0.2.0
geonamescache==1.0.2
gitdb2==2.0.5
GitPython==2.1.11
grpcio==1.12.1
h5py==2.8.0
html5lib==0.9999999
hyperas==0.4
hyperopt==0.1.1
idna==2.8
imageio==2.3.0
ipykernel==5.1.0
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.4
jedi==0.13.3
Jinja2==2.10.1
joblib==0.12.3
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keras-tqdm==2.0.1
kerasplotlib==0.1.4
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
librosa==0.6.1
llvmlite==0.24.0
mando==0.6.4
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.3
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.0.6
mkl-random==1.0.1
nbconvert==5.5.0
nbformat==4.4.0
networkx==2.2
nltk==3.4.5
notebook==5.7.8
numba==0.39.0
numpy==1.16.0
opencv-python==3.4.2.17
pandas==0.24.0
pandocfilters==1.4.2
parso==0.4.0
patsy==0.5.1
pbr==5.2.1
pep8==1.7.1
pep8-naming==0.8.2
pickleshare==0.7.5
Pillow==5.4.1
plotly==3.10.0
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.6.0
PyAudio==0.2.11
pycodestyle==2.5.0
pycodestyle-magic==0.2.5
pycparser==2.19
pydocstyle==3.0.0
pydot==1.2.4
pydub==0.22.1
pyflakes==2.1.1
Pygments==2.4.0
pylint==2.2.2
pyloudnorm==0.0.1
pymongo==3.7.2
pyOpenSSL==19.0.0
pyparsing==2.2.0
PySocks==1.7.0
PySoundFile==0.9.0.post1
python-dateutil==2.8.0
python-speech-features==0.6
pytz==2019.1
pywinpty==0.5.5
PyYAML==5.1
pyzmq==17.1.2
qtconsole==4.4.4
radon==3.0.3
requests==2.21.0
resampy==0.2.1
retrying==1.3.3
scikit-learn==0.20.2
scipy==1.1.0
seaborn==0.9.0
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
sklearn==0.0
smmap2==2.0.5
snowballstemmer==1.2.1
statsmodels==0.9.0
stevedore==1.30.1
talos==0.5.0
tensorboard==1.12.2
tensorflow==1.12.2
tensorflow-gpu==1.12.0
termcolor==1.1.0
terminado==0.8.2
testfixtures==6.8.2
testpath==0.4.2
tornado==5.1.1
tqdm==4.32.1
traitlets==4.3.2
typed-ast==1.1.0
typing-extensions==3.7.2
urllib3==1.24.2
wcwidth==0.1.7
wemake-python-styleguide==0.8.1
Werkzeug==0.15.3
widgetsnbextension==3.4.2
win-inet-pton==1.1.0
wincertstore==0.2
wrangle==0.6.6
wrapt==1.10.11
